update sha256 input, types
This commit parses the long bool array from the Rust API into a format
acceptable by the sha256 implementation.

It also switches the sha256 implementation to use signed integers
instead of my custom unsigned (requires not-patch to be applied).

I'm not sure if the signed int switch works in general. I need tests!!