update token interface (rust compat+new protocol)
The token API needed some changes for compatibility with rust. This
commit adds
- explicit struct types
- chars instead of bools
- no includes in token.h

Remaining compatability requirements are:
- convert strings to char *s
- discuss specific format of char arrays

We've also updated the protocol since the token.h API was last changed,
so I made substantial changes to the parameters to reflect the new
version of the protocol.

Both sets of changes required some updates to other functions to fully
propogate through the code.

This commit breaks ECDSA tests :(